Fix `pipsi list` when ~/.local is a symlink
`pipsi list` would produce an empty list when ~/.local was a symlink to
another directory. This was because the comparison between binary paths
and venv paths was using dereferenced paths for the binaries but
non-dereferenced paths for the venvs. There's already a perfectly-good
function in the code that does this operation correctly,
`Repo.find_installed_executables()`, so `Repo.list_everything()` can
just use that instead.